further cleaning up valuations.h
